Refactor dupe query functions
